projects/personal-knowledge-hub/requirements.txt
# Core dependencies
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Document processing
PyPDF2>=3.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
python-docx>=0.8.11
markdown>=3.4.3

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
numpy>=1.24.3

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0